+++ sigopt/cli/utils.py
@@ -4,7 +4,7 @@ import subprocess
 import sys
 import threading
 
-from ..config import config
+from ..run_context import GlobalRunContext
 from ..vendored import six
 
 
@@ -68,18 +68,18 @@ def get_git_hexsha():
   except InvalidGitRepositoryError:
     return None
 
-def get_subprocess_environment(env=None):
+def get_subprocess_environment(config, env=None):
   ret = os.environ.copy()
   ret.update(config.get_environment_context())
   ret.update(env or {})
   return ret
 
-def run_subprocess(run_context, entrypoint, entrypoint_args, env=None):
+def run_subprocess(config, run_context, entrypoint, entrypoint_args, env=None):
   cmd = [sys.executable, entrypoint] + list(entrypoint_args)
-  return run_subprocess_command(run_context, cmd=cmd, env=env)
+  return run_subprocess_command(config, run_context, cmd=cmd, env=env)
 
-def run_subprocess_command(run_context, cmd, env=None):
-  env = get_subprocess_environment(env)
+def run_subprocess_command(config, run_context, cmd, env=None):
+  env = get_subprocess_environment(config, env)
   proc_stdout, proc_stderr = subprocess.PIPE, subprocess.PIPE
   proc = subprocess.Popen(
     cmd,
@@ -107,7 +107,7 @@ def run_subprocess_command(run_context, cmd, env=None):
   if return_code > 0:
     raise subprocess.CalledProcessError(return_code, cmd)
 
-def run_notebook(entrypoint):
+def run_notebook(config, entrypoint):
   return subprocess.check_output(
     [
       'jupyter', 'nbconvert',
@@ -119,30 +119,31 @@ def run_notebook(entrypoint):
       '--to=python',
       entrypoint,
     ],
-    env=get_subprocess_environment(),
+    env=get_subprocess_environment(config),
   )
 
-def run_user_program(run_factory, entrypoint, entrypoint_args, user_provided_name=None, suggestion=None):
-  with run_factory.create_global_run(name=user_provided_name, suggestion=suggestion) as run_context:
-    if config.code_tracking_enabled:
-      source_code = {}
-      with open(entrypoint) as entrypoint_fp:
-        source_code['content'] = entrypoint_fp.read()
-        git_hash = get_git_hexsha()
-        if git_hash:
-          source_code['hash'] = git_hash
-      run_context.log_source_code(**source_code)
-    if entrypoint.endswith('.ipynb'):
-      if entrypoint_args:
-        raise Exception('Command line arguments cannot be passed to notebooks')
-      # TODO(patrick): The output of this command should be the tracked code (or logs? It's kind of both)
-      run_notebook(entrypoint)
-    else:
-      run_subprocess(run_context, entrypoint, entrypoint_args)
+def run_user_program(config, run_context, entrypoint, entrypoint_args, user_provided_name=None):
+  if config.code_tracking_enabled:
+    source_code = {}
+    with open(entrypoint) as entrypoint_fp:
+      source_code['content'] = entrypoint_fp.read()
+      git_hash = get_git_hexsha()
+      if git_hash:
+        source_code['hash'] = git_hash
+    run_context.log_source_code(**source_code)
+  global_run_context = GlobalRunContext(run_context)
+  config.set_context_entry(global_run_context)
+  if entrypoint.endswith('.ipynb'):
+    if entrypoint_args:
+      raise Exception('Command line arguments cannot be passed to notebooks')
+    # TODO(patrick): The output of this command should be the tracked code (or logs? It's kind of both)
+    run_notebook(config, entrypoint)
+  else:
+    run_subprocess(config, run_context, entrypoint, entrypoint_args)
 
 def check_path(entrypoint, error_msg):
   if not os.path.isfile(entrypoint):
     raise Exception(error_msg)
 
-def setup_cli():
+def setup_cli(config):
   config.set_user_agent_info(['CLI'])
