+++ sigopt/lib.py
@@ -56,3 +56,6 @@ def find(lis, predicate):
   Finds the first element in lis satisfying predicate, or else None
   """
   return next((item for item in lis if predicate(item)), None)
+
+def remove_nones(mapping):
+  return {key: value for key, value in mapping.items() if value is not None}
